backend/requirements.txt
# Core backend
fastapi==0.110.2
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.9.2

# Database (MySQL)
mysql-connector-python==9.0.0
sqlalchemy==2.0.35
alembic==1.13.2   # for DB migrations if needed
passlib[bcrypt]==1.7.4  # for password hashing

# Audio processing
pydub==0.25.1
soundfile==0.13.1
noisereduce==3.0.3

# NLP / ML
git+https://github.com/openai/whisper.git
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
scipy==1.12.0

# Filler words & text matching
rapidfuzz==3.14.1

# Utilities
tqdm==4.67.1
matplotlib==3.10.0
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic
passlib[bcrypt]
